+++ live_tests/rule_26_chain-conflict/015_rename_value_missing_type.sql
@@ -1,0 +1,1 @@
+ALTER TYPE missing_enum RENAME VALUE 'old' TO 'new';